Switch nav with personality
