+++ apps/web/src/components/Layout/LogoutButton.tsx
@@ -1,0 +1,24 @@
+import React from "react";
+
+interface LogoutButtonProps extends React.HTMLProps<HTMLButtonElement> {
+  children: React.ReactNode;
+}
+const LogoutButton = React.forwardRef<HTMLButtonElement, LogoutButtonProps>(
+  ({ children, ...rest }, forwardedRef) => (
+    <button
+      data-h2-color="base(black) base:hover(primary) base:iap(primary) base:iap:hover(primary.darker)"
+      data-h2-font-size="base(normal)"
+      data-h2-text-decoration="base(underline)"
+      style={{
+        background: "none",
+      }}
+      ref={forwardedRef}
+      {...rest}
+      type="button"
+    >
+      {children}
+    </button>
+  ),
+);
+
+export default LogoutButton;
